+++ src/migration/tests.rs
@@ -8,7 +8,7 @@ use giganto_client::ingest::network::{
 #[test]
 fn giganto_conn() {
     let data =
-        "1669735962.571151000	localhost	fe80::2267:7cff:fef0:cb09	133	ff02::2	134	1	0.000000000	-	0	0	1	0";
+        "1669735962.571151000	localhost	fe80::2267:7cff:fef0:cb09	133	ff02::2	134	1	sf	0.000000000	-	0	0	1	0";
 
     let rec = stringrecord(data);
 
@@ -17,7 +17,7 @@ fn giganto_conn() {
 
 #[test]
 fn giganto_http() {
-    let data = "1669773412.241856000	localhost	129.204.40.54	47697	218.144.35.150	80	0	0.000000000	GET	218.144.35.150	/root11.php	-	1.1	Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3440.106 Safari/537.36	0	286	302	Found	-	-	-	-	-	-	-	-	-	-";
+    let data = "1669773412.241856000	localhost	129.204.40.54	47697	218.144.35.150	80	0	0.000000000	GET	218.144.35.150	/root11.php	-	1.1	Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3440.106 Safari/537.36	0	286	302	Found	-	-	-	-	-	-	-	-	-	-	10,10,10	-";
 
     let rec = stringrecord(data);
 
@@ -36,7 +36,7 @@ fn giganto_rdp() {
 
 #[test]
 fn giganto_smtp() {
-    let data = "1669136154.131718000	localhost	220.73.219.213	51280	67.195.204.72	25	0	0.000000000	hanjinyea@monami.com	-	-	-	-	-";
+    let data = "1669136154.131718000	localhost	220.73.219.213	51280	67.195.204.72	25	0	0.000000000	hanjinyea@monami.com	-	-	-	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -54,7 +54,7 @@ fn giganto_dns() {
 
 #[test]
 fn giganto_ntlm() {
-    let data = "1614130258.669753000	localhost	10.200.90.100	59271	192.168.0.111	445	0	0.000000000	it	-	-	HYPER01	HYPER01.hansae.local	hansae.local	-";
+    let data = "1614130258.669753000	localhost	10.200.90.100	59271	192.168.0.111	445	0	0.000000000	-	it	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -72,7 +72,7 @@ fn giganto_kerberos() {
 
 #[test]
 fn giganto_ssh() {
-    let data = "1562093121.802019000	localhost	114.249.237.38	41260	203.254.132.18	22	0	0.000000000	2	F	1	-	SSH-2.0-Go	SSH-1.99-Cisco-1.25	aes128-cbc	hmac-sha1	none	diffie-hellman-group1-sha1	ssh-rsa	e8:1a:0c:fe:e5:d9:0d:d7:2f:9f:a2:c4:4a:36:c5:24";
+    let data = "1562093121.802019000	localhost	114.249.237.38	41260	203.254.132.18	22	0	0.000000000	SSH-2.0-Go	SSH-1.99-Cisco-1.25	aes128-cbc	hmac-sha1	none	diffie-hellman-group1-sha1	ssh-rsa	-	-	-	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -117,7 +117,7 @@ fn giganto_ldap() {
 
 #[test]
 fn giganto_tls() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	server_name	alpn_protocol	ja3	version	10	ja3s	serial	sub_country	sub_org_name	sub_comm_name	1	2	sub_alt_name	issuer_country	issuer_org_name	issuer_org_unit_name	issuer_common_name	10";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	server_name	alpn_protocol	ja3	version	771,769,770	0,1,2	10	0,1	ja3s	serial	sub_country	sub_org_name	sub_comm_name	1	2	sub_alt_name	issuer_country	issuer_org_name	issuer_org_unit_name	issuer_common_name	10";
 
     let rec = stringrecord(data);